GridPartition.processNearby implemented & ComponentRegistry intern logic changed

+++ src/main/java/World/Systems/System.java
@@ -4,11 +4,11 @@ import GameObject.GameObject;
 import java.util.List;
 
 public abstract class System {
-    protected final int REQUIRED_SIGNATURE;
+    protected final long REQUIRED_SIGNATURE;
     private final boolean THREADING;
     private float lastExecutionTimeMs; // Para el profiling
 
-    public System(int requiredSignature, boolean threading) {
+    public System(long requiredSignature, boolean threading) {
         this.REQUIRED_SIGNATURE = requiredSignature;
         this.THREADING = threading;
     }
